+++ app/core/config.py
@@ -53,6 +53,7 @@ class Settings(BaseSettings):
     KLAS_FILE_LIST_URL: str
     KLAS_LECTURE_URL: str
     KLAS_RECORDED_LECTURE_URL: str
+    KLAS_COURSE_INFO_URL: str
     KLAS_PROFILE_URL: str
     KLAS_STUDENT_INFO_URL: str
     KLAS_STUDENT_INFO_API_URL: str